+++ NEWS.md
@@ -1,7 +1,7 @@
 # *baRulho 1.0.7*
 
 * `spectral_correlation()`, `spcc()` and `spectral_blur_ratio()` renamed to `spectrum_correlation()`, `spcc_distortion()` and `spectrum_blur_ratio()` respectively
-* Fix bug in `spcc_distortion()`
+* Fix bug in `spcc_distortion()` and `excess_attenuation()`
 
 # *baRulho 1.0.6*